♻️ [refactor]: #452 Phase 0 bootstrap: lexer tests.rs に分離（内容変更なし）
lexer.rs の `#[cfg(test)] mod tests { ... }` ブロック 2912 行を
`#[cfg(test)] mod tests;` の 1 行宣言に置換し、旧 `mod tests` の
内側 (use super::*; + 293 件のテスト関数 + Phase 区切りコメント) を
src/lexer/tests.rs に退避した。

- 293 件すべてグリーン維持
- 短縮テスト関数名集合 diff 0 行 / 件数 293 / 重複 0
- cargo fmt / clippy グリーン
- lexer.rs: 3572 行 → 662 行

Refs #452
